refactor : Point, Segment 패키지 이동 (path -> node)
